chore: pin dredd
This change introduces explicit test every time Dredd is released. When we do the release, https://greenkeeper.io/ notices it and creates a Pull Request at this repository, which runs the tests and verified the just-released version of Dredd works correctly with the example. Dredd already uses this repo in its own smoke tests, but I think it is better to have the release train explicit and to make sure the CI configs work as intended and Dredd works as intended with all the CIs (e.g. with AppVeyor - Windows), with every release.